Make random identifiers available only in unit tests
