+++ pkg/sdk/random_test.go
@@ -1,0 +1,21 @@
+package sdk
+
+import (
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/helpers/random"
+)
+
+func randomSchemaObjectIdentifier() SchemaObjectIdentifier {
+	return NewSchemaObjectIdentifier(random.StringN(12), random.StringN(12), random.StringN(12))
+}
+
+func randomExternalObjectIdentifier() ExternalObjectIdentifier {
+	return NewExternalObjectIdentifier(NewAccountIdentifierFromAccountLocator(random.StringN(12)), randomAccountObjectIdentifier())
+}
+
+func randomDatabaseObjectIdentifier() DatabaseObjectIdentifier {
+	return NewDatabaseObjectIdentifier(random.StringN(12), random.StringN(12))
+}
+
+func randomAccountObjectIdentifier() AccountObjectIdentifier {
+	return NewAccountObjectIdentifier(random.StringN(12))
+}
